requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pillow-avif-plugin
jinja2